Filter elements with no id for Docsify Search
When using docsify search and clicking on a specific search result, the ToC does not get updated and we get an error in the console "appendChild" for the wrapper element on createList function.
This error was triggered because the plugin was pulling more elements than needed. The filter solves the issue.